+++ app/src/main/java/de/thomaskuenneth/benice/AppChooserActivity.kt
@@ -27,8 +27,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckBox
 import androidx.compose.material.icons.filled.CheckBoxOutlineBlank
-import androidx.compose.material.icons.filled.FilterAlt
-import androidx.compose.material.icons.filled.FilterAltOff
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
@@ -56,7 +54,6 @@ import androidx.window.core.layout.WindowSizeClass
 import androidx.window.core.layout.WindowWidthSizeClass
 import androidx.window.layout.WindowMetricsCalculator
 
-private const val PREFS_FILTER_ON = "filterOn"
 private const val PREFS_LAUNCH_ADJACENT = "launchAdjacent"
 
 class AppChooserActivity : ComponentActivity() {
@@ -72,7 +69,6 @@ class AppChooserActivity : ComponentActivity() {
         prefs = PreferenceManager.getDefaultSharedPreferences(this)
         enableEdgeToEdge()
         val viewModel by viewModels<BeNiceViewModel>()
-        viewModel.setFilterOn(prefs.getBoolean(PREFS_FILTER_ON, true))
         viewModel.setLaunchAdjacent(prefs.getBoolean(PREFS_LAUNCH_ADJACENT, true))
         windowSizeClass = computeWindowSizeClass()
         setContent {
@@ -88,25 +84,6 @@ class AppChooserActivity : ComponentActivity() {
                             title = { Text(text = stringResource(id = R.string.app_name)) },
                             scrollBehavior = scrollBehavior,
                             actions = {
-                                IconButtonWithTooltip(
-                                    onClick = {
-                                        val newValue = !state.filterOn
-                                        viewModel.setFilterOn(newValue)
-                                        prefs.edit().putBoolean(PREFS_FILTER_ON, newValue).apply()
-                                    },
-                                    imageVector = if (state.filterOn) {
-                                        Icons.Default.FilterAlt
-                                    } else {
-                                        Icons.Default.FilterAltOff
-                                    },
-                                    contentDescription = stringResource(
-                                        id = if (state.filterOn) {
-                                            R.string.filter_on
-                                        } else {
-                                            R.string.filter_off
-                                        }
-                                    )
-                                )
                                 val menuItems = mutableListOf<@Composable () -> Unit>()
                                 if (!windowSizeClass.hasExpandedScreen()) {
                                     menuItems.add {
@@ -220,7 +197,7 @@ class AppChooserActivity : ComponentActivity() {
         }
     }
 
-    private fun onAppsForAppPairSelected(firstApp: AppInfo, secondApp: AppInfo) {
+    private fun onAppsForAppPairSelected(firstApp: AppInfo, secondApp: AppInfo, delay: Long) {
         if (shortcutManager.isRequestPinShortcutSupported) {
             val id = "${firstApp.className}|${secondApp.className}"
             val label = "${firstApp.label} \u2011 ${secondApp.label}"
@@ -234,7 +211,7 @@ class AppChooserActivity : ComponentActivity() {
                     )
                 )
                 .setShortLabel(label)
-                .setIntent(createLaunchAppPairIntent(firstApp, secondApp))
+                .setIntent(createLaunchAppPairIntent(firstApp, secondApp, delay))
                 .build()
             shortcutManager.requestPinShortcut(shortcutInfo, null)
         }