+++ app/src/main/java/de/thomaskuenneth/benice/BeNiceScreen.kt
@@ -1,33 +1,33 @@
 package de.thomaskuenneth.benice
 
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.AddLink
 import androidx.compose.material.icons.filled.Create
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.Launch
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Surface
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Slider
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
@@ -35,13 +35,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import androidx.window.core.layout.WindowSizeClass
 import androidx.window.core.layout.WindowWidthSizeClass
 import kotlinx.coroutines.launch
@@ -54,7 +50,7 @@ fun BeNiceScreen(
     onClick: (AppInfo, Boolean) -> Unit,
     onAddLinkClicked: (AppInfo) -> Unit,
     onOpenAppInfoClicked: (AppInfo) -> Unit,
-    onAppsForAppPairSelected: (AppInfo, AppInfo) -> Unit,
+    onAppsForAppPairSelected: (AppInfo, AppInfo, Long) -> Unit,
     modifier: Modifier
 ) {
     var contextMenuAppInfo by remember { mutableStateOf<AppInfo?>(null) }
@@ -70,6 +66,7 @@ fun BeNiceScreen(
             callback()
         }
     }
+    var showAppPairDialog by remember { mutableStateOf(false) }
     Box(
         modifier = modifier,
         contentAlignment = Alignment.Center
@@ -77,18 +74,31 @@ fun BeNiceScreen(
     {
         when (state.isLoading) {
             true -> CircularProgressIndicator()
-            false -> AppChooser(
-                state = state,
-                columns = when (windowSizeClass.windowWidthSizeClass) {
-                    WindowWidthSizeClass.MEDIUM -> 2
-                    WindowWidthSizeClass.EXPANDED -> 3
-                    else -> 1
-                },
-                onClick = onClick,
-                onLongClick = { appInfo ->
-                    contextMenuAppInfo = appInfo
+            false -> {
+                AppChooser(
+                    installedApps = state.installedApps,
+                    columns = when (windowSizeClass.windowWidthSizeClass) {
+                        WindowWidthSizeClass.MEDIUM -> 2
+                        WindowWidthSizeClass.EXPANDED -> 3
+                        else -> 1
+                    },
+                    onClick = onClick,
+                    onLongClick = { appInfo ->
+                        contextMenuAppInfo = appInfo
+                    }
+                )
+                FloatingActionButton(
+                    onClick = { showAppPairDialog = true },
+                    modifier = Modifier
+                        .align(alignment = Alignment.BottomEnd)
+                        .padding(end = 16.dp, bottom = 16.dp)
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Add,
+                        contentDescription = stringResource(id = R.string.create_app_pair)
+                    )
                 }
-            )
+            }
         }
         contextMenuAppInfo?.let {
             ModalBottomSheet(
@@ -132,92 +142,94 @@ fun BeNiceScreen(
         }
     }
     if (showSelectSecondAppDialog) {
-        Dialog(
+        AppChooserDialog(
+            installedApps = state.installedApps,
+            onClick = { secondApp, _ ->
+                showSelectSecondAppDialog = false
+                firstApp?.let { onAppsForAppPairSelected(it, secondApp, 500L) }
+            },
             onDismissRequest = {
                 showSelectSecondAppDialog = false
             }
-        ) {
-            Surface(
-                color = MaterialTheme.colorScheme.surface,
-                shadowElevation = 4.dp,
-                shape = RoundedCornerShape(size = 12.dp)
-            ) {
-                AppChooser(
-                    state = state,
-                    columns = 1,
-                    onClick = { secondApp, _ ->
-                        showSelectSecondAppDialog = false
-                        firstApp?.let {
-                            onAppsForAppPairSelected(
-                                it,
-                                secondApp
-                            )
-                        }
-                    },
-                    onLongClick = {}
-                )
+        )
+    }
+    if (showAppPairDialog) {
+        AppPairDialog(
+            state = state,
+            onDismissRequest = { showAppPairDialog = false },
+            onFinished = { first, second, delay ->
+                onAppsForAppPairSelected(first, second, delay)
+                showAppPairDialog = false
             }
-        }
+        )
     }
 }
 
-@OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun AppChooser(
+fun AppPairDialog(
     state: BeNiceScreenUiState,
-    columns: Int,
-    onClick: (AppInfo, Boolean) -> Unit,
-    onLongClick: (AppInfo) -> Unit
+    onDismissRequest: () -> Unit,
+    onFinished: (AppInfo, AppInfo, Long) -> Unit
 ) {
-    val haptics = LocalHapticFeedback.current
-    if (state.installedApps.isEmpty()) {
-        Text(
-            modifier = Modifier
-                .width(600.dp)
-                .padding(horizontal = 16.dp),
-            text = stringResource(
-                id = if (state.filterOn) {
-                    R.string.no_portrait_apps
-                } else {
-                    R.string.no_apps
-                }
-            ),
-            style = MaterialTheme.typography.displaySmall,
-            color = MaterialTheme.colorScheme.primary,
-            textAlign = TextAlign.Start
-        )
-    } else {
-        LazyVerticalGrid(
-            modifier = Modifier.fillMaxSize(),
-            columns = GridCells.Fixed(count = columns)
-        ) {
-            state.installedApps.forEach { appInfo ->
-                item {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .combinedClickable(
-                                onClick = { onClick(appInfo, false) },
-                                onLongClick = {
-                                    haptics.performHapticFeedback(HapticFeedbackType.LongPress)
-                                    onLongClick(appInfo)
-                                },
-                                onLongClickLabel = stringResource(id = R.string.open_context_menu)
-                            )
-                            .padding(horizontal = 24.dp, vertical = 16.dp),
-                        horizontalArrangement = Arrangement.Start,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        AppIconImage(drawable = appInfo.icon)
-                        Spacer(modifier = Modifier.width(16.dp))
-                        Text(
-                            text = appInfo.label,
-                            style = MaterialTheme.typography.bodyLarge,
-                            color = MaterialTheme.colorScheme.primary
-                        )
-                    }
+    var firstApp: AppInfo? by remember { mutableStateOf(null) }
+    var secondApp: AppInfo? by remember { mutableStateOf(null) }
+    var delay by remember { mutableFloatStateOf(500F) }
+    AlertDialog(
+        onDismissRequest = onDismissRequest,
+        confirmButton = {
+            Button(
+                enabled = firstApp != null && secondApp != null,
+                onClick = { onFinished(firstApp!!, secondApp!!, delay.toLong()) }) {
+                Text(text = stringResource(id = R.string.create))
+            }
+        },
+        dismissButton = {
+            OutlinedButton(onClick = { onDismissRequest() }) {
+                Text(text = stringResource(id = R.string.cancel))
+            }
+        },
+        title = { Text(text = stringResource(id = R.string.create_app_pair)) },
+        text = {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                CompactAppChooser(
+                    installedApps = state.installedApps,
+                    selectedApp = firstApp,
+                    hint = R.string.first_app,
+                    onItemClicked = { selectedApp -> firstApp = selectedApp }
+                )
+                CompactAppChooser(
+                    installedApps = state.installedApps,
+                    selectedApp = secondApp,
+                    hint = R.string.second_app,
+                    onItemClicked = { selectedApp -> secondApp = selectedApp }
+                )
+                Text(
+                    modifier = Modifier.padding(top = 16.dp),
+                    text = stringResource(id = R.string.launch_after),
+                    style = MaterialTheme.typography.labelLarge
+                )
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    Text(
+                        text = stringResource(id = R.string.five_hundred_ms),
+                        style = MaterialTheme.typography.labelMedium
+                    )
+                    Slider(
+                        value = delay,
+                        onValueChange = { delay = it },
+                        valueRange = (500F..2000F),
+                        steps = 14,
+                        modifier = Modifier.weight(1.0F)
+                    )
+                    Text(
+                        text = stringResource(id = R.string.two_secs),
+                        style = MaterialTheme.typography.labelMedium
+                    )
                 }
             }
         }
-    }
+    )
 }