+++ app/src/main/java/de/thomaskuenneth/benice/BeNiceViewModel.kt
@@ -12,7 +12,6 @@ import kotlinx.coroutines.launch
 
 
 data class BeNiceScreenUiState(
-    val filterOn: Boolean = false,
     val launchAdjacent: Boolean = false,
     val isLoading: Boolean = false,
     val installedApps: List<AppInfo> = emptyList()
@@ -22,17 +21,6 @@ class BeNiceViewModel : ViewModel() {
 
     private val _uiState = MutableStateFlow(BeNiceScreenUiState())
     val uiState: StateFlow<BeNiceScreenUiState> = _uiState.asStateFlow()
-
-    private val installedAppsFlow: MutableStateFlow<List<AppInfo>> = MutableStateFlow(emptyList())
-
-    fun setFilterOn(filterOn: Boolean) {
-        _uiState.update { currentState ->
-            currentState.copy(
-                filterOn = filterOn
-            )
-        }
-        updateInstalledApps()
-    }
 
     fun setLaunchAdjacent(launchAdjacent: Boolean) {
         _uiState.update { currentState ->
@@ -45,8 +33,13 @@ class BeNiceViewModel : ViewModel() {
     fun queryInstalledApps(packageManager: PackageManager) {
         setLoading(true)
         viewModelScope.launch(Dispatchers.IO) {
-            installedAppsFlow.value = installedApps(packageManager)
-            updateInstalledApps()
+            _uiState.update { currentState ->
+                currentState.copy(
+                    installedApps = installedApps(
+                        packageManager = packageManager
+                    ).sortedBy { it.label }
+                )
+            }
             setLoading(false)
         }
     }
@@ -55,15 +48,6 @@ class BeNiceViewModel : ViewModel() {
         _uiState.update { currentState ->
             currentState.copy(
                 isLoading = isLoading
-            )
-        }
-    }
-
-    private fun updateInstalledApps() {
-        _uiState.update { currentState ->
-            currentState.copy(
-                installedApps = installedAppsFlow.value.filter { appInfo -> !currentState.filterOn || appInfo.screenOrientationPortrait }
-                    .sortedBy { it.label }
             )
         }
     }