+++ whatsnew.md
@@ -1,3 +1,5 @@
 # What's new
 
-- You can now create app pairs
+- Removed filter because it was less convenient than anticipated
+- Improved app pair creation
+- You can now change the delay for launching the second app